Fix warning.
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@4593 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ eklm/simulation/src/FiberAndElectronics.cc
@@ -256,8 +256,8 @@ void EKLM::FiberAndElectronics::simulateADC()
 {
   int i;
   for (i = 0; i < m_nTimeDigitizationSteps; i++) {
-    m_ADCAmplitude[i] = 0.5 * ADCRange *
-                        m_digitizedAmplitude->GetBinContent(i + 1);
+    m_ADCAmplitude[i] = (int)(0.5 * ADCRange *
+                              m_digitizedAmplitude->GetBinContent(i + 1));
   }
 }
